update default packer

+++ benchmarks_test.go
@@ -29,7 +29,7 @@ func Benchmark_NoRoute(b *testing.B) {
 		panic(err)
 	}
 	// defer client.Close() // nolint
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: []byte("ping")})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: []byte("ping")})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
@@ -41,7 +41,7 @@ func Benchmark_NotFoundHandler(b *testing.B) {
 		DoNotPrintRoutes: true,
 	})
 	s.NotFoundHandler(func(ctx *Context) (*message.Entry, error) {
-		return ctx.Response(uint32(0), []byte("not found"))
+		return ctx.Response(0, []byte("not found"))
 	})
 	go s.Serve(":0") // nolint
 
@@ -54,7 +54,7 @@ func Benchmark_NotFoundHandler(b *testing.B) {
 	}
 	// defer client.Close() // nolint
 
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: []byte("ping")})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: []byte("ping")})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
@@ -65,7 +65,7 @@ func Benchmark_OneHandler(b *testing.B) {
 	s := NewServer(&ServerOption{
 		DoNotPrintRoutes: true,
 	})
-	s.AddRoute(uint32(1), func(ctx *Context) (*message.Entry, error) {
+	s.AddRoute(1, func(ctx *Context) (*message.Entry, error) {
 		return ctx.Response(2, []byte("pong"))
 	})
 	go s.Serve(":0") // nolint
@@ -79,7 +79,7 @@ func Benchmark_OneHandler(b *testing.B) {
 	}
 	// defer client.Close() // nolint
 
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: []byte("ping")})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: []byte("ping")})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
@@ -97,8 +97,8 @@ func Benchmark_ManyHandlers(b *testing.B) {
 		}
 	}
 
-	s.AddRoute(uint32(1), func(ctx *Context) (*message.Entry, error) {
-		return ctx.Response(uint32(2), []byte("pong"))
+	s.AddRoute(1, func(ctx *Context) (*message.Entry, error) {
+		return ctx.Response(2, []byte("pong"))
 	}, m, m)
 
 	go s.Serve(":0") // nolint
@@ -112,7 +112,7 @@ func Benchmark_ManyHandlers(b *testing.B) {
 	}
 	// defer client.Close() // nolint
 
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: []byte("ping")})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: []byte("ping")})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
@@ -123,10 +123,10 @@ func Benchmark_OneRouteSet(b *testing.B) {
 	s := NewServer(&ServerOption{
 		DoNotPrintRoutes: true,
 	})
-	s.AddRoute(uint32(1), func(ctx *Context) (*message.Entry, error) {
+	s.AddRoute(1, func(ctx *Context) (*message.Entry, error) {
 		ctx.Set("key", "value")
 		v, _ := ctx.Get("key")
-		return ctx.Response(uint32(2), []byte(v.(string)))
+		return ctx.Response(2, []byte(v.(string)))
 	})
 	go s.Serve(":0") // nolint
 
@@ -139,7 +139,7 @@ func Benchmark_OneRouteSet(b *testing.B) {
 	}
 	// defer client.Close() // nolint
 
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: []byte("ping")})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: []byte("ping")})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
@@ -151,7 +151,7 @@ func Benchmark_OneRouteJsonCodec(b *testing.B) {
 		Codec:            &JsonCodec{},
 		DoNotPrintRoutes: true,
 	})
-	s.AddRoute(uint32(1), func(ctx *Context) (*message.Entry, error) {
+	s.AddRoute(1, func(ctx *Context) (*message.Entry, error) {
 		req := make(map[string]string)
 		if err := ctx.Bind(&req); err != nil {
 			panic(err)
@@ -169,7 +169,7 @@ func Benchmark_OneRouteJsonCodec(b *testing.B) {
 	}
 	// defer client.Close() // nolint
 
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: []byte(`{"data": "ping"}`)})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: []byte(`{"data": "ping"}`)})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
@@ -181,7 +181,7 @@ func Benchmark_OneRouteProtobufCodec(b *testing.B) {
 		Codec:            &ProtobufCodec{},
 		DoNotPrintRoutes: true,
 	})
-	s.AddRoute(uint32(1), func(ctx *Context) (*message.Entry, error) {
+	s.AddRoute(1, func(ctx *Context) (*message.Entry, error) {
 		var req pb.Sample
 		if err := ctx.Bind(&req); err != nil {
 			panic(err)
@@ -200,7 +200,7 @@ func Benchmark_OneRouteProtobufCodec(b *testing.B) {
 	// defer client.Close() // nolint
 
 	data, _ := s.Codec.Encode(&pb.Sample{Foo: "test", Bar: 1})
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: data})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: data})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
@@ -212,7 +212,7 @@ func Benchmark_OneRouteMsgpackCodec(b *testing.B) {
 		Codec:            &MsgpackCodec{},
 		DoNotPrintRoutes: true,
 	})
-	s.AddRoute(uint32(1), func(ctx *Context) (*message.Entry, error) {
+	s.AddRoute(1, func(ctx *Context) (*message.Entry, error) {
 		var req msgpack.Sample
 		if err := ctx.Bind(&req); err != nil {
 			panic(err)
@@ -231,7 +231,7 @@ func Benchmark_OneRouteMsgpackCodec(b *testing.B) {
 	// defer client.Close() // nolint
 
 	data, _ := s.Codec.Encode(&msgpack.Sample{Foo: "test", Bar: 1})
-	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: uint32(1), Data: data})
+	packedMsg, _ := s.Packer.Pack(&message.Entry{ID: 1, Data: data})
 	beforeBench(b)
 	for i := 0; i < b.N; i++ {
 		_, _ = client.Write(packedMsg)
